Rename c4r -> cloud4rpi
