+++ Makefile
@@ -7,7 +7,7 @@ style:
 	pep8 --show-source --config=.pep8_setup.cfg --show-pep8 .
 
 lint:
-	pylint --rcfile=.pylintrc --reports=n c4r/*.py examples/*.py test/*.py
+	pylint --rcfile=.pylintrc --reports=n cloud4rpi/*.py examples/*.py test/*.py
 
 test:
 	python -m unittest discover test
