+++ cloud4rpi/api_client.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from c4r import config
+from cloud4rpi import config
 
 import re
 import json
@@ -28,7 +28,7 @@ class MqttApi(object):
         def noop_on_command(cmd):
             pass
 
-        client_id = 'c4r-{0}'.format(device_token)
+        client_id = 'cloud4rpi-{0}'.format(device_token)
         # TODO: clean_session=False
         self.__client = mqtt.Client(client_id)
         self.__host = host
