+++ cloud4rpi/cpu_temperature.py
