+++ cloud4rpi/device.py
