+++ test/test_sensors.py
@@ -2,8 +2,8 @@
 
 import unittest
 import pyfakefs.fake_filesystem_unittest as ffut
-import c4r
-import c4r.ds18b20
+import cloud4rpi
+import cloud4rpi.ds18b20
 
 sensor_10 = \
     '2d 00 4d 46 ff ff 08 10 fe : crc=fe YES' '\n' \
@@ -30,26 +30,26 @@ class TestDs18b20Sensors(ffut.TestCase):
         )
 
     def testFindDSSensors(self):
-        sensors = c4r.DS18b20.find_all()
+        sensors = cloud4rpi.DS18b20.find_all()
         self.assertEqual(len(sensors), 2)
         self.assertEqual(sensors[0].address, '10-000802824e58')
         self.assertEqual(sensors[1].address, '28-000802824e58')
 
     def testRead(self):
-        sensor = c4r.DS18b20('28-000802824e58')
+        sensor = cloud4rpi.DS18b20('28-000802824e58')
         result = sensor.read()
         self.assertEqual(result, 28.250)
 
     def testRaisesExceptionOnInvalidAddress(self):
-        with self.assertRaises(c4r.ds18b20.InvalidW1Address):
-            c4r.DS18b20('invalid address')
+        with self.assertRaises(cloud4rpi.ds18b20.InvalidW1Address):
+            cloud4rpi.DS18b20('invalid address')
 
 
 @unittest.skip("Not tests actually")
 class TestNet(unittest.TestCase):
     def testTemporary(self):
-        host = c4r.Hostname()
-        ip = c4r.IPAddress()
+        host = cloud4rpi.Hostname()
+        ip = cloud4rpi.IPAddress()
 
         print(host.read())
         print(ip.read())
